+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Type a new commit message at the top of the list
 - Improve `fixup!` functions: create and apply
 - Reword the selected commit message with `r`
+- Add light theme: `-light` option and `l` hotkey to switch
 
 ## [0.2.0][] - 2026-09-22